Reworked into a scheduler, multiple files
