+++ src/cloud/benchflow/data-analyses-scheduler/scriptSubmission.go
@@ -1,0 +1,52 @@
+package main
+
+import (
+	"fmt"
+	"os/exec"
+	"os"
+	"strings"
+	"bytes"
+)
+
+// Function that submits a script with spark-submit and checks the output for errors
+func submitScript(args []string, script string) bool {
+	retries := 0
+	cmd := exec.Command(sparkHome+"/bin/spark-submit", args...)
+	for retries < 3 {
+		retries += 1
+		cmd.Stdout = os.Stdout
+		errOutput := &bytes.Buffer{}
+		cmd.Stderr = errOutput
+		err := cmd.Start()
+		cmd.Wait()
+		if err != nil {
+			panic(err)
+			}
+		errLog := errOutput.String()
+		fmt.Println(errLog)
+		if checkForErrors(errLog) {
+			fmt.Println("Script " + script + " failed")
+			fmt.Println(errLog)
+			continue
+		}
+		fmt.Println("Script "+script+" processed")
+		break
+	}
+	if retries == 3 {
+		fmt.Println("Max number of retries reached for " + script)
+		return false
+		}
+	return true
+	}
+
+// Checks if the output log for spark-submit contains errors
+func checkForErrors(errLog string) bool {
+	errString := strings.ToLower(errLog)
+	if strings.Contains(errString, "error") {
+		return true
+		}
+	if strings.Contains(errString, "exception") {
+		return true
+		}
+	return false
+	}
