+++ src/cloud/benchflow/data-analyses-scheduler/transformerWorker.go
@@ -1,0 +1,57 @@
+package main
+
+import (
+ 	"fmt"
+)
+
+// Create new worker
+func NewTransformerWorker(id int, workerQueue chan chan WorkRequest) TransformerWorker {
+  worker := TransformerWorker{
+    ID:          id,
+    Work:        make(chan WorkRequest),
+    WorkerQueue: workerQueue,
+    QuitChan:    make(chan bool)}
+  
+  return worker
+}
+
+// Worker struct
+type TransformerWorker struct {
+  ID          int
+  Work        chan WorkRequest
+  WorkerQueue chan chan WorkRequest
+  QuitChan    chan bool
+}
+
+// Start worker
+func (w *TransformerWorker) Start() {
+    go func() {
+      for {
+      	// Worker adds itself to the worker queue
+        w.WorkerQueue <- w.Work
+        
+        select {
+    	// Takes work from its own queue
+        case work := <-w.Work:
+          // Submits script to Spark
+          success := submitScript(work.SparkArgs, work.Script)
+          if success {
+			  meetRequirement(work.Topic, work.TrialID, work.ExperimentID, "trial")
+			  AnalyserDispatchRequestsQueue <- work
+		  }
+          
+        // Quits if signalled
+        case <-w.QuitChan:
+          fmt.Printf("worker%d stopping\n", w.ID)
+          return
+        }
+      }
+    }()
+}
+
+// Stop worker
+func (w *TransformerWorker) Stop() {
+  go func() {
+    w.QuitChan <- true
+  }()
+}
